refacto : raise more specific errors
